pyproject.toml
[tool.poetry]
name = "webm-dr"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.isort]
profile = "black"